adapted 1-click-farming for the last time
For technical reasons the first adaption of the 1-click-farming could not be
allowed by tribalwars (though it was compatible to the rules). In particular,
the bot detection mechanism would have reported false detections, eventually
leading to locked TW accounts (again).

To avoid these false detections, the rally point may not be skipped.

This commit updates the sendtroops functionality to fulfil this requirement.
In addition, it adds a userscript that users must install if they want the
rally point's form to be filled out automatically with the proper units and
target coordinates.