Farmmanager: Fix Farm-Assistant issue by properly escaping report data.
The Farm-Assistant is kind of a farmmanager provided by TribalWars itself. It can
be enabled ingame with premium points.

Players repeatedly reported that the Farm-Assistant, if enabled, modified the
report page, confusing NoPaste's report parsing engine.

The problem was caused by the Farm-Assistant inserting some javascript in the
report. The javascript contained some URL-creating code, containing the
character "&" which was not properly encoded in the POSTDATA sent to
NoPaste.